requirements.txt
Flask==2.3.3
numpy==1.24.3
Pillow==10.0.0
tensorflow==2.13.0
scikit-learn==1.3.0
pandas==2.0.3
joblib==1.3.2
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.1
opencv-python==4.8.0.76
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
APScheduler==3.10.4